inspect performance

+++ objects/video.php
@@ -850,12 +850,13 @@ if (!class_exists('Video')) {
 
         static function getAllVideosAsync($status = "viewable", $showOnlyLoggedUserVideos = false, $ignoreGroup = false, $videosArrayId = array(), $getStatistcs = false, $showUnlisted = false, $activeUsersOnly = true) {
             global $global;
+            $return = array();
             $users_id = User::getId();
             $get = json_encode(@$_GET);
             $post = json_encode(@$_POST);
             $md5 = md5("{$users_id}{$get}{$post}{$status}{$showOnlyLoggedUserVideos}{$ignoreGroup}" . implode("_", $videosArrayId) . "{$getStatistcs}{$showUnlisted}{$activeUsersOnly}");
             $cacheFileName = $global['systemRootPath'] . "videos/cache/getAllVideosAsync_{$md5}";
-            if (!file_exists($cacheFileName)) {
+            if (!file_exists($cacheFileName) && !file_exists($cacheFileName.".lock")) {
                 $total = static::getAllVideos($status, $showOnlyLoggedUserVideos, $ignoreGroup, $videosArrayId, $getStatistcs, $showUnlisted, $activeUsersOnly);
                 file_put_contents($cacheFileName, json_encode($total));
                 return $total;
